Remove Option wrapper from background colour definitions
This is because `None` in this situation represents transparency, but
there is already a way to represent transparency...!  I've just added
a few new ways to get to it.

I'll likely make it so that the internal representation for colours is
not CSS soon, to push all the floating point math out of the hot path.